+++ tests/providers/amazon/aws/operators/test_ecs.py
@@ -25,8 +25,8 @@ from copy import deepcopy
 import mock
 from parameterized import parameterized
 
-from airflow.contrib.operators.ecs_operator import ECSOperator
 from airflow.exceptions import AirflowException
+from airflow.providers.amazon.aws.operators.ecs import ECSOperator
 
 RESPONSE_WITHOUT_FAILURES = {
     "failures": [],
@@ -52,7 +52,7 @@ RESPONSE_WITHOUT_FAILURES = {
 
 class TestECSOperator(unittest.TestCase):
 
-    @mock.patch('airflow.contrib.operators.ecs_operator.AwsHook')
+    @mock.patch('airflow.providers.amazon.aws.operators.ecs.AwsHook')
     def setUp(self, aws_hook_mock):
         self.aws_hook_mock = aws_hook_mock
         self.ecs_operator_args = {
@@ -96,7 +96,7 @@ class TestECSOperator(unittest.TestCase):
     ])
     @mock.patch.object(ECSOperator, '_wait_for_task_ended')
     @mock.patch.object(ECSOperator, '_check_success_task')
-    @mock.patch('airflow.contrib.operators.ecs_operator.AwsHook')
+    @mock.patch('airflow.providers.amazon.aws.operators.ecs.AwsHook')
     def test_execute_without_failures(self, launch_type, tags, aws_hook_mock,
                                       check_mock, wait_mock):
         client_mock = aws_hook_mock.return_value.get_client_type.return_value
